Working Index.vue page.
Changed the package.json dependencies to have a working match with inertia and vue. All index.vue components work as intended and the movies are displayed on the screen and filters are reactive and work as intended. Cleaned up unneeded dependencies in the package.json, node modules and package.lock