Api testing, Show.vue and front end styling fixes.
Fixed the frontend styling to be consistent across all pages and be responsive on all screen sizes. Implemented a show.vue page with full functionality. The user is now able to view the details of a single movie with edit and delete options listed. Finalized Unit and feature testing for the backend and API. All tests pass and the project is fully functional.