Seeder and pagination.
Implemented more movies into the database seeder. Fully tested the project found that pagination and the drop down menus do not work as intended. Fixed the page pagination by implementing buttons and fixed the display of the drop down menus for directors and year of release. Fully tested the project. Now works fully as intended.